particles update, mouse bug fix
git-svn-id: https://happy-engine.googlecode.com/svn/trunk@180 5aa1a52c-6ec2-ce28-c747-3d7cbe9a755d

+++ src/HappyEngine/code/graphics/Shader.cpp
@@ -84,7 +84,7 @@ bool validateProgram(GLuint programID)
         if (length > 0)
         {
             char sId[5];
-            sprintf(sId, "%d", programID);
+            sprintf(sId, "%d", (int)programID);
             HE_ERROR("Program " + std::string(sId) + "link err:");
             HE_ERROR(std::string(buffer));
         }
@@ -97,7 +97,7 @@ bool validateProgram(GLuint programID)
     if (validateStatus == GL_FALSE)
     {
         char sId[5];
-        sprintf(sId, "%d", programID);
+        sprintf(sId, "%d", (int)programID);
         HE_ERROR("Error validating shader " + std::string(sId));
         succes = false;
     }
